I was testing this example code, and I found this error
Usually, for a fast testing I copy and paste the Code. 
I was testing this in an aplication that I am building, and when I run this code whit the 'Debug JS Remotly' this works correcty, but i quit the debug mode and the app no works, and I complete the arrow function and the app works fine.
Plaeas is a simple mistake but i loss two hours discovering the error.